requirements.txt
selenium
pytest
pytest-html
pytest-repeat
allure-pytest